pyproject.toml
[project]
name = "agent-poc-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "openai-agents>=0.14.1",
    "pydantic>=2.13.1",
    "python-dotenv>=1.2.2",
]
